Remove unnecessary casts
git-svn-id: https://svn.apache.org/repos/asf/commons/sandbox/compress/trunk@749856 13f79535-47bb-0310-9956-ffa450edef68

+++ src/test/java/org/apache/commons/compress/archivers/zip/ZipArchiveOutputStreamTest.java
@@ -66,7 +66,7 @@ public class ZipArchiveOutputStreamTest extends TestCase {
     }
 
     public void testAdjustToLong() {
-        assertEquals((long) Integer.MAX_VALUE,
+        assertEquals(Integer.MAX_VALUE,
                      ZipArchiveOutputStream.adjustToLong(Integer.MAX_VALUE));
         assertEquals(((long) Integer.MAX_VALUE) + 1,
                      ZipArchiveOutputStream.adjustToLong(Integer.MAX_VALUE + 1));
